2019 day 9 prep: i64 compute + relative mode

+++ src/puzzles_2019/day_05.rs
@@ -3,8 +3,8 @@ use crate::puzzle_input;
 
 pub fn run() {
     if let Ok(input) = puzzle_input::read_string("./input/2019-d05-input1.txt") {
-        let mut codes: Vec<i32> = input.trim().split(',')
-                                    .map(|x| x.trim().parse::<i32>().unwrap())
+        let mut codes: Vec<i64> = input.trim().split(',')
+                                    .map(|x| x.trim().parse::<i64>().unwrap())
                                     .collect();
 
         let mut io = compute::DefaultProgramIO::new(vec![1]);
@@ -13,8 +13,8 @@ pub fn run() {
     }
 
     if let Ok(input) = puzzle_input::read_string("./input/2019-d05-input1.txt") {
-        let mut codes: Vec<i32> = input.trim().split(',')
-                                    .map(|x| x.trim().parse::<i32>().unwrap())
+        let mut codes: Vec<i64> = input.trim().split(',')
+                                    .map(|x| x.trim().parse::<i64>().unwrap())
                                     .collect();
 
         let mut io = compute::DefaultProgramIO::new(vec![5]);
